eurocargo_management/backend/requirements.txt
dotenv
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic[email]
pydantic-settings==2.7.0
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
bcrypt>=4.0
fpdf2>=2.7
python-jose[cryptography]==3.3.0

# Test dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
